Move MIME types to separate modules
