+++ webgear-core/src/WebGear/Core/Trait/Auth/Common.hs
@@ -21,11 +21,12 @@ import Data.Void (absurd)
 import GHC.TypeLits (KnownSymbol, Symbol, symbolVal)
 import Web.HttpApiData (FromHttpApiData (..))
 import WebGear.Core.Handler (Handler, unwitnessA, (>->))
+import WebGear.Core.MIMEType.PlainText (PlainText)
 import WebGear.Core.Modifiers (Existence (..), ParseStyle (..))
 import WebGear.Core.Request (Request)
 import WebGear.Core.Response (Response)
 import WebGear.Core.Trait (Get (..), Sets, With)
-import WebGear.Core.Trait.Body (Body, PlainText, setBody)
+import WebGear.Core.Trait.Body (Body, setBody)
 import WebGear.Core.Trait.Header (RequestHeader (..), RequiredResponseHeader, setHeader)
 import WebGear.Core.Trait.Status (Status, unauthorized401)
 import Prelude hiding (break, drop)
